Refactor query execution components: unify diagnostics observer references, enhance README documentation, and improve code clarity across various executors and tests.
