+++ src/Shardis.Query.InMemory/README.md
@@ -1,9 +1,25 @@
 # Shardis.Query.InMemory
 
-In-memory query executor for prototyping and unit tests.
+An in-memory query executor for Shardis used in tests and examples. This package provides a lightweight `IQueryExecutor` implementation that yields deterministic results and is useful in unit tests and local samples.
 
-## Features
+## When to use
 
-- Executes delegates across in-memory shard collections
-- Unordered and ordered streaming merge testing
-- Deterministic scheduling via test helpers
+- Use in tests, examples and demos where you need deterministic, fast in-memory shard-local query behavior.
+
+## What the package provides
+
+- `InMemoryQueryExecutor` implementation and small helpers for seeding test data.
+
+## Links
+
+- Tests: `test/Shardis.Query.Tests`
+
+## Quick usage example
+
+```csharp
+var exec = new InMemoryQueryExecutor(...); // create with seeded test data
+await foreach (var item in exec.QueryAsync(...))
+{
+    // assert test expectations
+}
+```
