+++ src/Shardis.Query/IShardQueryCapabilities.cs
@@ -8,10 +8,3 @@ public interface IShardQueryCapabilities
     /// <summary>True if pagination (Skip/Take) is supported.</summary>
     bool SupportsPagination { get; }
 }
-
-internal sealed class BasicQueryCapabilities(bool ordering, bool pagination) : IShardQueryCapabilities
-{
-    public bool SupportsOrdering { get; } = ordering;
-    public bool SupportsPagination { get; } = pagination;
-    public static readonly IShardQueryCapabilities None = new BasicQueryCapabilities(false, false);
-}
