+++ test/Shardis.Migration.Tests/Executor_SmokeTests.cs
@@ -1,14 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
-using Shardis.Migration;
+
 using Shardis.Migration.Abstractions;
 using Shardis.Migration.Execution;
 using Shardis.Migration.Model;
 using Shardis.Model;
-using Xunit;
 
 namespace Shardis.Migration.Tests;
